remove ComposeUITextField.uikit.kt
Moved to UIKitRenderSync demo.
Also copied to example with interop https://github.com/JetBrains/compose-multiplatform/pull/3582/commits/c7f06d07429c940d3809012b2e2daf5a19a271df
