+++ compose/mpp/demo/src/uikitMain/kotlin/bugs/UIKitRenderSync.kt
@@ -16,27 +16,25 @@
 
 package bugs
 
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.material.Text
 import androidx.compose.material.TextField
 import androidx.compose.mpp.demo.Screen
+import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
+import androidx.compose.runtime.rememberUpdatedState
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.interop.ComposeUITextField
 import androidx.compose.ui.interop.UIKitView
-import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.unit.dp
+import kotlinx.cinterop.ObjCAction
 import kotlinx.cinterop.readValue
-import platform.CoreGraphics.CGRect
+import platform.CoreGraphics.CGRectMake
 import platform.CoreGraphics.CGRectZero
+import platform.Foundation.NSSelectorFromString
 import platform.UIKit.*
 
 val UIKitRenderSync = Screen.Example("UIKitRenderSync") {
@@ -54,9 +52,47 @@ val UIKitRenderSync = Screen.Example("UIKitRenderSync") {
                     },
                     modifier = Modifier.fillMaxWidth().height(40.dp)
                 )
-                2 -> TextField(text, onValueChange = { text = it}, Modifier.fillMaxWidth())
-                else -> ComposeUITextField(text, onValueChange = { text = it }, Modifier.fillMaxWidth().height(40.dp))
+
+                2 -> TextField(text, onValueChange = { text = it }, Modifier.fillMaxWidth())
+                else -> ComposeUITextField(
+                    text,
+                    onValueChange = { text = it },
+                    Modifier.fillMaxWidth().height(40.dp)
+                )
             }
         }
     }
+}
+
+/**
+ * Compose wrapper for native UITextField.
+ * @param value the input text to be shown in the text field.
+ * @param onValueChange the callback that is triggered when the input service updates the text. An
+ * updated text comes as a parameter of the callback
+ * @param modifier a [Modifier] for this text field. Size should be specified in modifier.
+ */
+@Composable
+private fun ComposeUITextField(value: String, onValueChange: (String) -> Unit, modifier: Modifier) {
+    val latestOnValueChanged by rememberUpdatedState(onValueChange)
+
+    UIKitView(
+        factory = {
+            val textField = object : UITextField(CGRectMake(0.0, 0.0, 0.0, 0.0)) {
+                @ObjCAction
+                fun editingChanged() {
+                    latestOnValueChanged(text ?: "")
+                }
+            }
+            textField.addTarget(
+                target = textField,
+                action = NSSelectorFromString(textField::editingChanged.name),
+                forControlEvents = UIControlEventEditingChanged
+            )
+            textField
+        },
+        modifier = modifier,
+        update = { textField ->
+            textField.text = value
+        }
+    )
 }
